--- Editor/ZedDiscovery.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 269d046c929614d7dad1cb25b6419a45
